Restructured to a more common structure
Renamed Controller directory to Api.
Moved the task to its own scheduledtasks directory.
