+++ Jellyfin.Plugin.Template/Api/JellybenchApiController.cs
